Improve display of research essays and entries
* adjust spacing

* wrap a tag

* improve command structure

* support for png

* no longer float images

* update inline essay links

* add search help

* adjust line height

* pull entries from right place

* update docs

* display categories

* add annotations to entry list

* don't link to items that are ready.

* guard against bad links

* italics for categories

* adjust quick search tips

* fix header prob

* adjust annotation heading

+++ scripts/lizard.js
@@ -41,6 +41,7 @@ const docxMimeType = "application/vnd.openxmlformats-officedocument.wordprocessi
 const spreadsheetMimeType = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet";
 const rCloneExportFormats = "docx,csv";
 const jpegMimeType = "image/jpeg";
+const pngMimeType = "image/png";
 const googleLinkRegX = /https:\/\/drive\.google\.com\/open\?id=/;
 const googleLinkRegX2 = /https:\/\/drive.google.com\/file\/d\//;
 const videoEmbedRegX = /^https:\/\/academiccommons\.columbia\.edu/;
@@ -345,7 +346,7 @@ function locateAnnotationAssets() {
                             // locate the illustrations, if any
                             if( assetFile.children.length > 0 && assetFile.name.includes('Illustrations_') ) {
                                 assetFile.children.forEach( illustrationFile => {
-                                    if( illustrationFile.mimeType === jpegMimeType ) {
+                                    if( illustrationFile.mimeType === jpegMimeType || illustrationFile.mimeType === pngMimeType ) {
                                         illustrations.push( illustrationFile );
                                     }
                                 });
@@ -368,7 +369,7 @@ function locateAnnotationAssets() {
                         illustrations: illustrations
                     });
                     const path = nodeToPath(textFileNode);
-                    logger.info(`Found annotation: ${textFileNode.id} in ${googleShareName}${path}`);
+                    // logger.info(`Found annotation: ${textFileNode.id} in ${googleShareName}${path}`);
                 } else {
                     const path = nodeToPath(annotationRoot);
                     logger.info(`Annotation not found. Must contain docx file with Text_* in the filename: ${path}`);
@@ -388,7 +389,6 @@ function filterForDownload(annotationMetadata, annotationAssets) {
         const metadata = annotationMetadata[annotationAsset.id]
         if( metadata ) {        
             const {status,refresh} = metadata
-            const annotationDir = `${baseDir}/${annotationAsset.id}`;
 
             if( publicationStage === 'production') {
                 // download everything that is published, to make sure we have the latest
@@ -396,8 +396,8 @@ function filterForDownload(annotationMetadata, annotationAssets) {
                     selectedAssets.push(annotationAsset)
                 }
             } else {
-                // if user requests refresh or we don't have it yet and it is marked for publication
-                if(refresh || (!fs.existsSync(annotationDir) && (status === 'published' || status === 'staging')) ) {
+                // if user requests refresh 
+                if(refresh) {
                     selectedAssets.push(annotationAsset)
                 }
             }
@@ -427,7 +427,7 @@ function filterForPublication(annotationMetadata, annotationAssets) {
             // must be downloaded and marked for publication
             if( fs.existsSync(annotationDir) ) {
                 if( status === 'published' ) {
-                    selectedAssets[annotationAsset.id] = annotationAsset
+                    publishedAssets[annotationAsset.id] = annotationAsset
                     publishedMetadata[metadata.id] = metadata
                 } else {
                     // delete assets not heading to production
@@ -442,7 +442,9 @@ function filterForPublication(annotationMetadata, annotationAssets) {
 function deleteAnnotation(annotationID) {
     const annotationHTMLFile = `${targetAnnotationDir}/${annotationID}.html`;    
     const illustrationsDir = `${targetImageDir}/${annotationID}`;
-    fs.unlinkSync(annotationHTMLFile);
+    if( fs.existsSync(annotationHTMLFile) ) {
+        fs.unlinkSync(annotationHTMLFile);
+    }
     // TODO delete contents of illustration dir
 }
 
@@ -532,7 +534,8 @@ function syncDriveAssets( driveAssets ) {
         driveAsset.illustrations.forEach( illustration => {
             const illustrationSrc = `${googleShareName}${nodeToPath(illustration)}`;
             const illustrationTmp = `${illustrationsDir}/${illustration.name}`;
-            const illustrationDest = `${illustrationsDir}/${illustration.id}.jpg`;   
+            const ext = illustration.mimeType === jpegMimeType ? 'jpg' : 'png';
+            const illustrationDest = `${illustrationsDir}/${illustration.id}.${ext}`;  
             syncDriveFile(illustrationSrc, illustrationsDir);
             fs.renameSync(illustrationTmp,illustrationDest);
             illustrations.push(illustrationDest);
@@ -601,7 +604,7 @@ async function processAnnotations(annotationAssets, annotationMetadata, authors,
         let annotation
         if( asset ) {
             // if we have the asset, but it isn't marked for publication, just publish metadata
-            if( metadata.status !== 'staging' && metadata.status != 'production') {
+            if( metadata.status !== 'staging' && metadata.status != 'published') {
                 annotation = {
                     ...metadata,
                     annotationAuthors,
@@ -708,16 +711,19 @@ async function processAnnotation( annotationAsset, metadata, authors ) {
     }
     
     // Make a directory for the illustrations and copy them to there
+    const illustrations = {};
     const illustrationsDir = `${targetImageDir}/${annotationID}`;
     dirExists( illustrationsDir );
     annotationAsset.illustrations.forEach( illustration => {
         const sourceFile = `${baseDir}/${annotationAsset.id}/illustrations/${illustration}`
         const targetFile = `${illustrationsDir}/${illustration}`
+        const imageID = illustration.replace(/.\w+$/,'')
+        illustrations[imageID] = illustration
         fs.copyFileSync( sourceFile, targetFile );
     })
 
     // Take the pandoc output and transform it into final annotation html
-    processAnnotationHTML(annotationHTMLFile, annotationID, captions, biblio, altTexts);
+    processAnnotationHTML(annotationHTMLFile, annotationID, captions, biblio, illustrations, altTexts);
 
     return {
         ...metadata,
@@ -759,7 +765,7 @@ function findImageID( url ) {
     return null;
 }
 
-function processAnnotationHTML( annotationHTMLFile, annotationID, captions, biblio, altTexts ) {
+function processAnnotationHTML( annotationHTMLFile, annotationID, captions, biblio, illustrations, altTexts ) {
 
     logger.info(`Processing annotation ${annotationID}`);
     // load document 
@@ -776,6 +782,9 @@ function processAnnotationHTML( annotationHTMLFile, annotationID, captions, bibl
         const imageID = findImageID( href );
         const videoURL = !imageID && href.match(videoEmbedRegX) ? href : null;
         if( imageID || videoURL ) {
+            if( imageID && !illustrations[imageID] ) {
+                logger.info(`Illustration not found: ${href}`)
+            }
             const paragraphElement = findParentParagraph(anchorTag);     
             if( paragraphElement ) {
                 const figureRefEl = doc.createElement('i');
@@ -789,7 +798,7 @@ function processAnnotationHTML( annotationHTMLFile, annotationID, captions, bibl
                     if( !caption ) logger.info(`Caption not found for ${figureNumber}`);
                     const figCaption = (caption) ? `<figcaption>${caption}</figcaption>` : '';
                     if( imageID ) {
-                        const imageURL = `${imageRootURL}/${annotationID}/${imageID}.jpg`
+                        const imageURL = `${imageRootURL}/${annotationID}/${illustrations[imageID]}`
                         const altText = altTexts[imageID] ? altTexts[imageID].text : figureNumber
                         figureEl.innerHTML = `<img src="${imageURL}" alt="${altText}" />${figCaption}`;      
                     } else {
@@ -903,13 +912,22 @@ function setupLogging() {
 
 async function run(mode) {
     switch( mode ) {
+        case 'download-all': {
+            const annotationDriveAssets = locateAnnotationAssets();
+            syncDriveAssets( annotationDriveAssets );
+            }
+            break;
+        case 'download-thumbs': {
+            const thumbnailAssets = locateThumbnails();
+            const annotationMetadata = await loadAnnotationMetadata()
+            syncThumbnails( thumbnailAssets, annotationMetadata )
+            }
+            break;
         case 'download': {
             const annotationDriveAssets = locateAnnotationAssets();
-            const thumbnailAssets = locateThumbnails();
             const annotationMetadata = await loadAnnotationMetadata()
             const selectedAssets = filterForDownload(annotationMetadata,annotationDriveAssets)
             syncDriveAssets( selectedAssets );
-            syncThumbnails( thumbnailAssets, annotationMetadata )
             }
             break;
         case 'process': {
@@ -924,6 +942,12 @@ async function run(mode) {
         case 'index':
             // TODO only index the annotations published at this stage
             searchIndex.generateAnnotationIndex(targetAnnotationDir, targetSearchIndexDir);
+            break;
+        case 'init':
+            await run('download-all')
+            await run('download-thumbs')
+            await run('process')
+            await run('index')
             break;
         case 'run': 
             await run('download')
@@ -1000,10 +1024,13 @@ function main() {
     if( mode === 'help' ) {
         console.log(`Usage: lizard.js <command>` );
         console.log("A helpful lizard that responds to the following commands:")
-        console.log("\tdownload: Download the annotations from Google Drive via rclone.");
+        console.log("\tdownload-thumbs: Download all the essay thumbnails from Google Drive via rclone.");
+        console.log("\tdownload-all: Download all essays from Google Drive.");
+        console.log("\tdownload: Download only essays marked with as 'refresh'.");
         console.log("\tprocess: Process the downloaded files and place them on the asset server.");
-        console.log("\tindex: Create a search index of the annotations.");
-        console.log("\trun: Do all of the above.")
+        console.log("\tindex: Create a search index of the essays.");
+        console.log("\trun: Download, process, and index.")
+        console.log("\tinit: Download all, download thumbs, process, and index.")
         console.log("\thelp: Displays this help. ");
         process.exit(-1);
     }
